pyproject.toml
[tool.poetry]
name = "traducao-artigos-azure-ai"
version = "0.1.0"
description = ""
authors = ["Raul Gomes <rsgomes86@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.6.1"
python-docx = "^1.1.2"
streamlit = "^1.40.2"
streamlit-lottie = "^0.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
